Fix UDP_GRO struct cmsghdr data type

+++ examples/client.cc
@@ -909,7 +909,7 @@ int Client::on_read(const Endpoint &ep) {
   msg.msg_iov = &msg_iov;
   msg.msg_iovlen = 1;
 
-  uint8_t msg_ctrl[CMSG_SPACE(sizeof(int)) + CMSG_SPACE(sizeof(uint16_t))];
+  uint8_t msg_ctrl[CMSG_SPACE(sizeof(int)) + CMSG_SPACE(sizeof(int))];
   msg.msg_control = msg_ctrl;
 
   for (;;) {
